alliance flow and refactor folder structure
