+++ src/server/api/modules/alliance/dtos/responses/alliance-create.response.dto.ts
