+++ src/server/api/modules/alliance/dtos/responses/alliance-delete.response.dto.ts
