+++ src/server/api/modules/alliance/exceptions/alliance-not-found.exception.ts
@@ -1,5 +1,5 @@
 import { NOT_FOUND } from 'http-status-codes';
-import { HttpException } from '../../../common/exceptions/http.exception';
+import { HttpException } from '../../../../common/exceptions/http.exception';
 
 export class AllianceNotFoundException extends HttpException {
   constructor(id: string) {
