+++ src/server/api/modules/alliance/interfaces/responses/alliance-create.response.interface.ts
@@ -1,4 +1,4 @@
-import { Response } from '../../../../common/interfaces/response-object.interface';
+import { Response } from '../../../../../common/interfaces/response-object.interface';
 import { AllianceCreateResponseDto } from '../../dtos/responses/alliance-create.response.dto';
 
 export interface AllianceCreateResponse extends Response {
