+++ src/server/api/modules/alliance/interfaces/responses/alliance-delete.response.interface.ts
@@ -1,4 +1,4 @@
-import { Response } from '../../../../common/interfaces/response-object.interface';
+import { Response } from '../../../../../common/interfaces/response-object.interface';
 import { AllianceDeleteResponseDto } from '../../dtos/responses/alliance-delete.response.dto';
 
 export interface AllianceDeleteResponse extends Response {
