+++ src/server/api/modules/alliance/interfaces/responses/alliance-get.response.interface.ts
@@ -1,4 +1,4 @@
-import { Response } from '../../../../common/interfaces/response-object.interface';
+import { Response } from '../../../../../common/interfaces/response-object.interface';
 import { AllianceGetResponseDto } from '../../dtos/responses/alliance-get.response.dto';
 
 export interface AllianceGetResponse extends Response {
