+++ src/server/api/modules/alliance/interfaces/responses/alliance-list.response.interface.ts
@@ -1,4 +1,4 @@
-import { Response } from '../../../../common/interfaces/response-object.interface';
+import { Response } from '../../../../../common/interfaces/response-object.interface';
 import { AllianceResponseDto } from '../../dtos/responses/alliance-response.dto';
 
 export interface AllianceListResponse extends Response {
