+++ src/server/api/modules/alliance/interfaces/responses/alliance-update.response.interface.ts
@@ -1,5 +1,5 @@
 import { AllianceUpdateResponseDto } from './../../dtos/responses/alliance-update.response.dto';
-import { Response } from '../../../../common/interfaces/response-object.interface';
+import { Response } from '../../../../../common/interfaces/response-object.interface';
 
 export interface AllianceUpdateResponse extends Response {
   data?: AllianceUpdateResponseDto;
